DEV: Convert component to a glimmer component
Why this change?

We are moving to glimmer components as a default in core so it makes
sense to start porting theme components to glimmer components as well.
Glimmer components can also be imported into other files making it
possible to pass the components directly to plugin APIs in the future.